+++ R/cross-val.R
@@ -164,7 +164,7 @@ ll_sdmTMB <- function(object, withheld_y, withheld_mu) {
 #'
 #' m_cv <- sdmTMB_cv(
 #'   density ~ 0 + depth_scaled + depth_scaled2,
-#'   data = pcod, mesh = mesh,
+#'   data = pcod, mesh = mesh, spatial = "off",
 #'   family = tweedie(link = "log"), k_folds = 2
 #' )
 #'